started a combat system
the enemy moves towards the player, prepares or winds up for an attack, then attacks - it playes an animation, the attack is read off weapon contacts, wich are turned on while the attack plays and off at the end. use distance atm which might be a perfomance drain. also restructured some code, so that abstract class Actor is the base of the player and the hostile enemy ai  so i can set a bool on either from stats to do damage effects