Added a complete New Rig Named Astrid
This rig has less bones than the old one. This rig has attatchable Weapons. Ive made a Club and a Wooden Sword. They are inside the same folder as > Objects >Characters > Astrid. To Attatch the weapon , you have to drag and drop it on top of Astrid when in Hierachy like you would an animator controller. The swords deform and follow with the animation of the Rig.